extract server status publishing into its own file
